backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
